+++ internalUsages/FeatureCheck.Domain/Projections/ClientLoyaltyPointLists/ClientLoyaltyPointQueryNext.cs
@@ -1,30 +1,40 @@
+using ResultBoxes;
 using Sekiban.Core.Query.MultiProjections;
 using Sekiban.Core.Query.QueryModel;
 using System.Collections.Immutable;
 namespace FeatureCheck.Domain.Projections.ClientLoyaltyPointLists;
 
-public class ClientLoyaltyPointQuery : IMultiProjectionListQuery<ClientLoyaltyPointListProjection, ClientLoyaltyPointQuery.Parameter,
-    ClientLoyaltyPointQuery_Response>
+public record ClientLoyaltyPointQueryNext(
+    Guid? BranchId,
+    Guid? ClientId,
+    int? PageSize,
+    int? PageNumber,
+    ClientLoyaltyPointQueryNext.FilterSortKey? SortKey1,
+    ClientLoyaltyPointQueryNext.FilterSortKey? SortKey2,
+    bool? SortKey1Asc,
+    bool? SortKey2Asc) : INextMultiProjectionListWithPagingQuery<ClientLoyaltyPointListProjection, ClientLoyaltyPointQuery_Response>
 {
     public enum FilterSortKey
     {
         BranchName, ClientName
     }
 
-    public IEnumerable<ClientLoyaltyPointQuery_Response> HandleSort(Parameter param, IEnumerable<ClientLoyaltyPointQuery_Response> filteredList)
+    public ResultBox<IEnumerable<ClientLoyaltyPointQuery_Response>> HandleSort(
+        IEnumerable<ClientLoyaltyPointQuery_Response> filteredList,
+        IQueryContext context)
     {
         var sort = new Dictionary<FilterSortKey, bool>();
-        if (param.SortKey1 != null)
+        if (SortKey1 != null)
         {
-            sort.Add(param.SortKey1.Value, param.SortKey1Asc ?? true);
+            sort.Add(SortKey1.Value, SortKey1Asc ?? true);
         }
-        if (param.SortKey2 != null)
+        if (SortKey2 != null)
         {
-            sort.Add(param.SortKey2.Value, param.SortKey2Asc ?? true);
+            sort.Add(SortKey2.Value, SortKey2Asc ?? true);
         }
         if (sort.Count == 0)
         {
-            return filteredList.OrderBy(m => m.BranchName).ThenBy(m => m.ClientName);
+            return ResultBox.FromValue(filteredList.OrderBy(m => m.BranchName).ThenBy(m => m.ClientName).AsEnumerable());
         }
         var result = filteredList;
         foreach (var (sortKey, index) in sort.Select((item, index) => (item, index)))
@@ -65,33 +75,23 @@ public class ClientLoyaltyPointQuery : IMultiProjectionListQuery<ClientLoyaltyPo
                         });
             }
         }
-        return result;
+        return ResultBox.FromValue(result);
     }
 
-    public IEnumerable<ClientLoyaltyPointQuery_Response> HandleFilter(
-        Parameter param,
-        MultiProjectionState<ClientLoyaltyPointListProjection> projection)
+    public ResultBox<IEnumerable<ClientLoyaltyPointQuery_Response>> HandleFilter(
+        MultiProjectionState<ClientLoyaltyPointListProjection> projection,
+        IQueryContext context)
     {
         var result = projection.Payload.Records.Select(
             m => new ClientLoyaltyPointQuery_Response(m.BranchId, m.BranchName, m.ClientId, m.ClientName, m.Point));
-        if (param.BranchId.HasValue)
+        if (BranchId.HasValue)
         {
-            result = result.Where(x => x.BranchId == param.BranchId.Value).ToImmutableList();
+            result = result.Where(x => x.BranchId == BranchId.Value).ToImmutableList();
         }
-        if (param.ClientId.HasValue)
+        if (ClientId.HasValue)
         {
-            result = result.Where(x => x.ClientId == param.ClientId.Value).ToImmutableList();
+            result = result.Where(x => x.ClientId == ClientId.Value).ToImmutableList();
         }
-        return result;
+        return ResultBox.FromValue(result);
     }
-
-    public record Parameter(
-        Guid? BranchId,
-        Guid? ClientId,
-        int? PageSize,
-        int? PageNumber,
-        FilterSortKey? SortKey1,
-        FilterSortKey? SortKey2,
-        bool? SortKey1Asc,
-        bool? SortKey2Asc) : IListQueryPagingParameter<ClientLoyaltyPointQuery_Response>;
 }
