+++ internalUsages/FeatureCheck.Domain/Projections/GeneralListQuerySample_Response.cs
@@ -1,0 +1,4 @@
+using Sekiban.Core.Query.QueryModel;
+namespace FeatureCheck.Domain.Projections;
+
+public record GeneralListQuerySample_Response(string Name, string BranchName) : IQueryResponse;
