+++ src/Sekiban.Core/Query/QueryModel/INextAggregateListQuery.cs
@@ -9,8 +9,3 @@ public interface INextAggregateListQuery<TAggregatePayload, TOutput> : INextAggr
     public ResultBox<IEnumerable<TOutput>> HandleFilter(IEnumerable<AggregateState<TAggregatePayload>> list, IQueryContext context);
     public ResultBox<IEnumerable<TOutput>> HandleSort(IEnumerable<TOutput> filteredList, IQueryContext context);
 }
-public interface INextGeneralListQuery<TOutput> : INextGeneralQueryCommon<TOutput>, INextListQueryCommon<TOutput> where TOutput : notnull
-{
-    public ResultBox<IEnumerable<TOutput>> HandleFilter(IQueryContext context);
-    public ResultBox<IEnumerable<TOutput>> HandleSort(IEnumerable<TOutput> filteredList, IQueryContext context);
-}
