+++ src/Sekiban.Core/Query/QueryModel/INextAggregateListQueryWithPaging.cs
@@ -3,4 +3,3 @@ namespace Sekiban.Core.Query.QueryModel;
 
 public interface INextAggregateListQueryWithPaging<TAggregatePayload, TOutput> : INextAggregateListQuery<TAggregatePayload, TOutput>,
     IQueryPagingParameterCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
-public interface INextGeneralListQueryWithPaging<TOutput> : INextGeneralListQuery<TOutput>, IQueryPagingParameterCommon where TOutput : notnull;
