+++ src/Sekiban.Core/Query/QueryModel/INextAggregateListQueryWithPagingAsync.cs
@@ -3,5 +3,3 @@ namespace Sekiban.Core.Query.QueryModel;
 
 public interface INextAggregateListQueryWithPagingAsync<TAggregatePayload, TOutput> : INextAggregateListQueryAsync<TAggregatePayload, TOutput>,
     IQueryPagingParameterCommon where TOutput : notnull where TAggregatePayload : IAggregatePayloadCommon;
-public interface INextGeneralListQueryWithPagingAsync<TOutput> : INextGeneralListQueryAsync<TOutput>, IQueryPagingParameterCommon
-    where TOutput : notnull;
