+++ src/Sekiban.Core/Query/QueryModel/INextGeneralListQueryWithPaging.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface INextGeneralListQueryWithPaging<TOutput> : INextGeneralListQuery<TOutput>, IQueryPagingParameterCommon where TOutput : notnull;
