+++ src/Sekiban.Core/Query/QueryModel/INextGeneralListQueryWithPagingAsync.cs
@@ -1,0 +1,4 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface INextGeneralListQueryWithPagingAsync<TOutput> : INextGeneralListQueryAsync<TOutput>, IQueryPagingParameterCommon
+    where TOutput : notnull;
