+++ src/Sekiban.Core/Query/QueryModel/INextGeneralQuery.cs
@@ -1,0 +1,7 @@
+using ResultBoxes;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface INextGeneralQuery<TOutput> : INextGeneralQueryCommon<TOutput>, INextQueryCommon<TOutput> where TOutput : notnull
+{
+    public ResultBox<TOutput> HandleFilter(IQueryContext context);
+}
