+++ src/Sekiban.Core/Query/QueryModel/INextGeneralQueryAsync.cs
@@ -1,0 +1,8 @@
+using ResultBoxes;
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface INextGeneralQueryAsync< TOutput> : INextGeneralQueryCommon< TOutput>,
+    INextQueryCommon<TOutput>, INextQueryAsyncGeneral where TOutput : notnull
+{
+    public Task<ResultBox<TOutput>> HandleFilterAsync( IQueryContext context);
+}
