+++ src/Sekiban.Core/Query/QueryModel/INextGeneralQueryCommon.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Core.Query.QueryModel;
+
+public interface INextGeneralQueryCommon<TOutput> where TOutput : notnull;
